app/src/main/java/gr/uth/cardshop/domain/Address.java
package gr.uth.cardshop.domain;

public class Address {
    private String address;
    private boolean isSelected;
    private String docId;

    public Address() {
    }

    public boolean isSelected() {
        return isSelected;
    }

    public void setSelected(boolean selected) {
        isSelected = selected;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getDocId() {
        return docId;
    }

    public void setDocId(String docId) {
        this.docId = docId;
    }
}
